# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a  Demo Template API with Swagger documentation",
        "title": "Demo Template",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/orders": {
            "get": {
                "description": "List orders and return an array of order data with purchase details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "List orders",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Skip records",
                        "name": "skip",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit records",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orders displayed",
                        "schema": {
                            "$ref": "#/definitions/handler.meta"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new order and return the order data with purchase details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "parameters": [
                    {
                        "description": "Create order request",
                        "name": "createOrderRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.createOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order created",
                        "schema": {
                            "$ref": "#/definitions/handler.orderResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "409": {
                        "description": "Data conflict error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "description": "Get an order by id and return the order data with purchase details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get an order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order displayed",
                        "schema": {
                            "$ref": "#/definitions/handler.orderResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.PaymentType": {
            "type": "string",
            "enum": [
                "CASH",
                "E-WALLET",
                "EDC"
            ],
            "x-enum-varnames": [
                "Cash",
                "EWallet",
                "EDC"
            ]
        },
        "handler.categoryResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Foods"
                }
            }
        },
        "handler.createOrderRequest": {
            "type": "object",
            "required": [
                "customer_name",
                "payment_id",
                "products",
                "total_paid"
            ],
            "properties": {
                "customer_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "payment_id": {
                    "type": "integer",
                    "example": 1
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.orderProductRequest"
                    }
                },
                "total_paid": {
                    "type": "integer",
                    "example": 100000
                }
            }
        },
...
        "handler.errorValidResponse": {
            "type": "object",
            "properties": {
                "errorno": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Error message"
                    ]
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "handler.meta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "skip": {
                    "type": "integer",
                    "example": 0
                },
                "total": {
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "handler.orderProductRequest": {
            "type": "object",
            "required": [
                "product_id",
                "qty"
            ],
            "properties": {
                "product_id": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "qty": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "handler.orderProductResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "1970-01-01T00:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "order_id": {
                    "type": "integer",
                    "example": 1
                },
                "price": {
                    "type": "number",
                    "example": 100000
                },
                "product": {
                    "$ref": "#/definitions/handler.productResponse"
                },
                "product_id": {
                    "type": "integer",
                    "example": 1
                },
                "qty": {
                    "type": "integer",
                    "example": 1
                },
                "total_final_price": {
                    "type": "number",
                    "example": 100000
                },
                "total_normal_price": {
                    "type": "number",
                    "example": 100000
                },
                "updated_at": {
                    "type": "string",
                    "example": "1970-01-01T00:00:00Z"
                }
            }
        },
        "handler.orderResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "1970-01-01T00:00:00Z"
                },
                "customer_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "payment_type": {
                    "$ref": "#/definitions/handler.paymentResponse"
                },
                "payment_type_id": {
                    "type": "integer",
                    "example": 1
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.orderProductResponse"
                    }
                },
                "receipt_id": {
                    "type": "string",
                    "example": "4979cf6e-d215-4ff8-9d0d-b3e99bcc7750"
                },
                "total_paid": {
                    "type": "number",
                    "example": 100000
                },
                "total_price": {
                    "type": "number",
                    "example": 100000
                },
                "total_return": {
                    "type": "number",
                    "example": 0
                },
                "updated_at": {
                    "type": "string",
                    "example": "1970-01-01T00:00:00Z"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "handler.paymentResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "logo": {
                    "type": "string",
                    "example": "https://example.com/cash.png"
                },
                "name": {
                    "type": "string",
                    "example": "Tunai"
                },
                "type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.PaymentType"
                        }
                    ],
                    "example": "CASH"
                }
            }
        },
        "handler.productResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/handler.categoryResponse"
                },
                "created_at": {
                    "type": "string",
                    "example": "1970-01-01T00:00:00Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image": {
                    "type": "string",
                    "example": "https://example.com/chiki-ball.png"
                },
                "name": {
                    "type": "string",
                    "example": "Chiki Ball"
                },
                "price": {
                    "type": "number",
                    "example": 5000
                },
                "sku": {
                    "type": "string",
                    "example": "9a4c25d3-9786-492c-b084-85cb75c1ee3e"
                },
                "stock": {
                    "type": "integer",
                    "example": 100
                },
                "updated_at": {
                    "type": "string",
                    "example": "1970-01-01T00:00:00Z"
                }
            }
        },
...
    }
}